+++ src/Shell/Shell.tsx
@@ -1,24 +1,7 @@
-import { MdOutlineAddPhotoAlternate } from 'react-icons/md';
 import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels';
-import { findImages } from '../core/file-manager';
 import MetadataEditor from '../Editor/MetadataEditor';
 import ImageGrid from '../ImageGrid/ImageGrid';
-
-function Toolbar() {
-  return (
-    <div className="bg-base-200 px-4 py-1">
-      <button
-        type="button"
-        className="btn btn-ghost"
-        // TODO: need to handle if findImages fails
-        onClick={() => findImages()}
-      >
-        <MdOutlineAddPhotoAlternate size={26} />
-        <span className="sr-only">Add Image</span>
-      </button>
-    </div>
-  );
-}
+import Toolbar from './Toolbar';
 
 function Shell() {
   return (
