+++ src/Shell/Toolbar.tsx
@@ -1,0 +1,20 @@
+import { MdOutlineAddPhotoAlternate } from 'react-icons/md';
+import { findImages } from '../core/file-manager';
+
+function Toolbar() {
+  return (
+    <div className="bg-base-200 px-4 py-1">
+      <button
+        type="button"
+        className="btn btn-ghost"
+        // TODO: need to handle if findImages fails
+        onClick={() => findImages()}
+      >
+        <MdOutlineAddPhotoAlternate size={26} />
+        <span className="sr-only">Add Image</span>
+      </button>
+    </div>
+  );
+}
+
+export default Toolbar;
